requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.2
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
py-bcrypt==0.4
pyasn1==0.5.0
pycparser==2.21
pyodbc==4.0.39
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.17
SQLAlchemy-SQLSchema==0.1
typing_extensions==4.7.1
urllib3==2.0.4
validate-docbr==1.10.0
websocket-client==1.6.2
Werkzeug==2.3.6
WTForms==3.0.1
